convert to sub/pub arch
now clients will subscribe to some streamer.
if the streamer doesn't yet exist, it will be spawned using the client provided configuration. and the client will be registered with as a listener to that streamer (so every new event from that streamer will be sent to that client).
on the other hand, if the streamer already exists and a client is trying to add/spawn it, the client will simply be registered to the alraedy runnning streamer and no changes to the streamer configuration is made (even if the clinet's wanted config is different). now new events from this streamer will also be sent to the newly added client.

a client can shut/(stop listening to) a streamer, in this case, if the streamer has no more clients to server, it will be terminated. otherwise, it's kept alive to sever other clients.

a clinet can also terminate the sse connection (remove_client), in which case all of its streamers are shut (same as the shut point above)