move the arc<mutex< from the controller one level up
this will allow us to reason about what feilds must be locked together for atomicity of streaming manager operations.

this also gets rid of the smart channel receiver that removes a channel from the controller when it drops. it's now the responsibilty of who called `create_channel` to call `remove_channel` when the channel is no longer used (that is, the responsibility of the streaming manager)