ensure atomicity in streaming manager operations
This refactor makes sure that we perform write operations on the streaming manager atomically.

We were previouly locking some fields of the manager to edit them then unlock them and lock another field. This clearly isn't atomic since there is a depedency between the data stored in the two fields and they should be set in one step with no operation in between.

One solution could have been to just wrap the whole manager with a mutex and not allow concurrent access to it. This would hurt the read perf though, which is critical for frequent send calls and broadcast calls.
Another would have been introducing a new field (mutex) called `critical` or something and lock it when we want no other writing method to execute in parallel. This is the approach taken here, but since `clinets` field is already locked with every operation (every operation mutating the state of the streaming manager must involve a client anyway: added, removed, newly listening, etc...), we now lock the `clients` field before any cirtical section till the end so we are sure no other mutating code will run in parallel.