Add fee estimation to the streaming API
I went with removing the fee estimation context and rpc methods but I
fount that it is used by mm2 itself and not just for API usage.

Maybe with some refactor we could get the mm2 usage decoupled from the
API one, and abandon the fee context (it looks possible from the
streaming API side).