broadcast all from the data asker
since there is no such a streamer id (the id was made up on the spot in the data asker) the client who should receive this message won't get it (not any client in fact). it seems like we can't somehow know the client, so sending the request to all clients instead.