remove eth gas estimator context
rely only on the streaming version

`gas_fee_estimator` was removed from coins file and is now set while enabling the streaming for the first time. One can choose to enable it in 'simiple' or 'provider' mode.
provider mode uses 'gas_api' defined in mm2.json, if not found we always fail back to simple again.