add tx (history) streaming
this adds tx history streaming for tx history supporting coins (except zcoin).
the streamer is fairly simple, just a forwarder that doesn't do any computation. The tx details are grapped from the tx history state machine and forwared to the streamer right before getting stored in the db.

To be able to access the `event_stream_manager` from the tx history statemachine, `MmArc` was made accessible through the coins (`MmCoin`). Chose the coins here and not directly embedding the MmArc inside the history statemachine since we will probably use the event streamer more and more in different areas of the code, so having it distributed among coins would make adding new streaming functionalies easier.
We could have also gone for cloning `event_stream_manager` to the coins instead of passing the `ctx`. Went for passing the ctx here since eth & tindermint already pass the ctx, so to avoid redundency and unify coins structure.