partial impl of zcoin tx history
this is still missing the translation from txhash to json tx details (this need to be fetched from the db).

this commit also reactivates the z balance streaming by using the streaming manager to send the balance change notification to the z balance event streamer (instead of sending the notification on the channel which no body was listening to after balance streaming refactor).