setting a default to client_id
this sets a default for the client id from SSE side (to zero) and from rpc enablers (tasks coming soon) as well.
also a new composition struct is added to the rpc enablers to stop the repitition of client_id in each streaming enabler rpc.
same technique should be followed with the task manager rpcs