set default cliend_id for rpc tasks
this sets a default client id to zero for rpc managed tasks, so for the default case of single user they don't have to provide an ID to get progress updates.
this also introduces a new composition struct `RpcInitReq` for init rpcs which that holds the clinet_id instead of having to declare it in each RPC request. also the client_id is not removed from the RpcTask and supplied to the task manager directly so we don't have to store the clinet_id in each task as well.